[refator] notiType으로 모듈화
