+++ src/utils/notiPushMessage.js
@@ -1,14 +1,6 @@
 const { Strings } = require('./strings');
-
+const { NotiType } = require('./notiType');
 require('dotenv').config({ path: '../.env' });
-
-const notiType = {
-  RESTOCK: '재입고',
-  OPEN_DAY: '오픈일',
-  PREORDER_CLOSE: '프리오더 마감',
-  SALE_START: '세일 시작',
-  SALE_CLOSE: '세일 마감',
-};
 
 const message = {
   notification: {
@@ -23,14 +15,14 @@ const message = {
 };
 
 function dataMessage(itemNotiType, itemId, deviceFcmToken) {
-  message.notification.body = `${notiType.itemNotiType} ${Strings.notiMessageDescription}`;
+  message.notification.body = `${NotiType.itemNotiType} ${Strings.notiMessageDescription}`;
   message.data.itemId = String(itemId);
   message.token = deviceFcmToken;
   return message;
 }
 
 function dataMessageWithCount(itemNotiType, itemId, notiCount, deviceFcmToken) {
-  message.notification.body = `${notiType.itemNotiType} 알림 외 ${notiCount}개의 ${Strings.notiMessageDescription}`;
+  message.notification.body = `${NotiType.itemNotiType} 알림 외 ${notiCount}개의 ${Strings.notiMessageDescription}`;
   message.data.itemId = String(itemId);
   message.token = deviceFcmToken;
   return message;
