+++ src/utils/notiType.js
@@ -1,0 +1,11 @@
+const notiType = {
+  RESTOCK: '재입고',
+  OPEN_DAY: '오픈일',
+  PREORDER_CLOSE: '프리오더 마감',
+  SALE_START: '세일 시작',
+  SALE_CLOSE: '세일 마감',
+};
+
+module.exports = {
+  notiType,
+};
